refactor(3120): split producer_fault unit tests to a sibling — fixes the file-size ratchet
The write-support gating attributes took producer_fault.rs to 821 lines, over the
~800 source campsite target, so the lite gate's file-size ratchet FAILed (grew an
over-threshold file). Lifted the inline test module VERBATIM into
producer_fault_tests.rs (#1116/#1135), the same treatment from_readers.rs got:
601 + 231 lines, both under target. No CQLITE_ALLOW_FILE_GROWTH.

Bodies unchanged; the module is still a CHILD of producer_fault via #[path], so
super::* resolves as before. Re-verified: minimal-build succeeds, reduced-feature
LIB test binary builds, all 9 producer_fault tests pass, cqlite-core --lib 3395
pass, clippy --all-targets --all-features clean.

Refs #3120
